pyproject.toml
[project]
name = "genobear"
version = "0.1.1"
description = "Genomic Database Management and Annotation Tool"
readme = "README.md"
authors = [
    { name = "antonkulaga", email = "antonkulaga@gmail.com" }
]
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.2",
    "eliot>=1.17.5",
    "fsspec>=2025.10.0",
    "huggingface-hub>=1.1.5",
    "pipefunc[all]>=0.88.0",
    "polars-bio>=0.18.0",
    "pooch>=1.8.2",
    "pycomfort>=0.0.18",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "tenacity>=9.1.2",
    "typer>=0.20.0"
]

[project.scripts]
#genobear = "genobear:app"
prepare = "genobear.prepare:app"
annotate = "genobear.annotate:app"
modules = "genobear.modules:app"
convert-vortex = "genobear.convert_vortex:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/genobear"]

[tool.hatch.metadata]
allow-direct-references = true

[dependency-groups]
dev = [
    "jupyterlab>=4.5.0",
    "marimo>=0.18.0",
    "pytest>=9.0.1",
    "python-lsp-server[all]>=1.13.2",
    "google-genai>=1.52.0",
    "ruff>=0.14.6",
    "sqlglot>=28.0.0",
    "vegafusion>=2.0.3",
    "vl-convert-python>=1.8.0",
    "vortex-data>=0.56.0",
]

[tool.pytest.ini_options]
markers = [
    "large_download: marks tests that download large files (multi-GB) - skip by default",
    "integration: marks integration tests that require network access",
    "slow: marks tests that take a long time to run"
]

# Default pytest behavior: extra verbose and no capture to allow Eliot/stdout output
addopts = "-vvv -s -m 'not large_download'"
testpaths = ["tests"]


[tool.marimo.runtime]
dotenv = [".env"]
